Implement auto discovery streaming feature and results page
- Added streaming support for auto discovery process.
- Created new views for handling auto discovery and results.
- Updated templates for displaying discovery progress and results.
- Enhanced the changelog with new features and fixes.
- Updated nginx configuration to support streaming.
- Adjusted requirements to include necessary packages for streaming.
