+++ homelab_operator/templates/html/auto_discover_results.html
@@ -1,0 +1,127 @@
+{% extends 'html/base.html' %}
+
+{% if homelab %}
+    {% block title %}
+        {{ homelab.name }} - Auto Discover Results
+    {% endblock %}
+{% endif %}
+
+{% block content %}
+    <div>
+        <span style="display: flex;">
+            <h2 style="margin-bottom: .25rem;">Auto Discover Results</h2>
+        </span>
+        <span class="soft">
+            This page shows the results of the auto discovery process. It lists all servers found in the specified network range.
+            You can use this to quickly identify devices and services running on your network.
+        </span>
+    </div>
+    <div class="spacer"></div>
+    {% if servers %}
+        <form id="discovery_form" method="post" action="/auto_discover_results/{{ network_id }}/">
+            {% csrf_token %}
+            <div class="tile">
+                <div style="display: flex;">
+                    <div>
+                        <h3 style="display: flex; margin-bottom: 0;">
+                            Discovered Servers
+                        </h3>
+                        <span class="soft">
+                            Select and edit servers you want to add to your homelab.
+                        </span>
+                    </div>
+                </div>
+                <div class="spacer"></div>
+                <div style="border-top: 1px solid #ccc;"></div>
+                <div class="spacer"></div>
+                {% for server in servers %}
+                    {% with id='server_'|add:server.ip_address %}
+                        <div style="display: flex;">
+                            <div>
+                                {% include 'widgets/checkbox.html' with id=id checked=True %}
+                            </div>
+                            <div style="margin: 0 auto auto .5rem;">
+                                <label for="{{ id }}_name">
+                                    <span class="soft">Server Name</span>
+                                </label>
+                                <div>
+                                    <input type="text" name="{{ id }}_name" value="{{ server.hostname }}" style="width: min-content;" id="{{ id }}_name" placeholder="Server Name">
+                                    <input type="hidden" name="{{ id }}_ip" value="{{ server.ip_address }}">
+                                    <input type="hidden" name="{{ id }}_mac" value="{{ server.mac_address }}">
+                                </div>
+                                <div>
+                                    <span class="soft">{{ server.ip_address }}</span>
+                                    <span class="soft">at</span>
+                                    <span class="soft">{{ server.mac_address }}</span>
+                                </div>
+                                <div class="soft">
+                                    {{ server.services|length }} {{ server.services|length|pluralize:"service,services" }} found on this IP.
+                                </div>
+                            </div>
+                        </div>
+                    {% endwith %}
+                {% endfor %}
+            </div>
+            <div class="spacer"></div>
+            <div class="tile">
+                <div style="display: flex;">
+                    <div>
+                        <h3 style="display: flex;">
+                            Discovered Services
+                        </h3>
+                        <span class="soft">
+                            Select and edit services.
+                        </span>
+                    </div>
+                </div>
+                <div class="spacer"></div>
+                <div style="border-top: 1px solid #ccc;"></div>
+                <div class="spacer"></div>
+                {% for server in servers %}
+                    {% for service in server.services %}
+                        {% with id='service_'|add:service.endpoint|add:'_'|add:service.name %}
+                            <div style="display: flex;">
+                                <div>
+                                    {% include 'widgets/checkbox.html' with id=id checked=True %}
+                                </div>
+                                <div style="margin: 0 auto auto .5rem;">
+                                    <label for="{{ id }}_name">
+                                        <span class="soft">Service Name</span>
+                                    </label>
+                                    <div>
+                                        <input type="text" name="{{ id }}_name" value="{{ service.name }}" style="width: min-content;" id="{{ id }}_name" placeholder="Service Name">
+                                        <input type="hidden" name="{{ id }}_endpoint" value="{{ service.endpoint }}">
+                                        <input type="hidden" name="{{ id }}_port" value="{{ service.port }}">
+                                        <input type="hidden" name="{{ id }}_url" value="{{ service.url }}">
+                                        <input type="hidden" name="{{ id }}_server_name" value="{{ server.hostname }}">
+                                    </div>
+                                    <div class="soft">on {{ server.hostname }} ({{ service.endpoint }})</div>
+                                </div>
+                            </div>
+                        {% endwith %}
+                    {% endfor %}
+                {% endfor %}
+            </div>
+            <div class="spacer"></div>
+            <div class="login">
+                <label for="homelab_select">
+                    <span class="soft">Select Homelab</span>
+                </label>
+                <select id="homelab_select" name="homelab_id">
+                    {% for homelab in homelabs %}
+                        <option value="{{ homelab.id }}" {% if user.profile.last_selected_homelab.id == homelab.id %}selected{% endif %}>{{ homelab.name }}</option>
+                    {% endfor %}
+                </select>
+                <input type="submit" value="Save Discovery Results"></input>
+                <div class="soft" style="text-align: center;">This will save the selected servers and services to your homelab.</div>
+            </div>
+        </form>
+    {% else %}
+        <div>
+            <b style="color: #666;">No servers were found.</b>
+            <div class="soft">
+                No servers were found in the specified network range. Please check your network settings and try again.
+            </div>
+        </div>
+    {% endif %}
+{% endblock %}
