+++ nginx.conf
@@ -49,6 +49,11 @@ http {
             proxy_pass http://localhost:8000/cron/;
         }
 
+        location /auto_discover_stream/ {
+            proxy_pass http://localhost:8000;
+            proxy_buffering off;
+        }
+
         location / {
             proxy_set_header Host $host;
             proxy_set_header X-Real-IP $remote_addr;
