+++ requirements.txt
@@ -1,4 +1,6 @@
 Django >= 5.2.1
+gunicorn >= 22.0.0
+gevent >= 24.2.1
 waitress >= 3.0.2
 Werkzeug >= 3.1.3
 psycopg >= 3.2.9
